backend/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.23.1
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
google-generativeai==0.8.5
sentence-transformers==2.7.0
scikit-learn==1.3.2
fastapi-limiter==0.1.6
redis==5.0.1
pydantic==2.7.4
